Split out mocks that actually aren't mocks
